requirements.txt
fastapi
uvicorn[standard]
sqlalchemy>=2.0
pymysql
python-dotenv
pydantic
python-multipart
pdfplumber
python-docx
requests

